requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.19.4
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sql==0.4.0
Werkzeug==1.0.1
